+++ CHANGELOG.rst
@@ -12,6 +12,9 @@ Unreleased
 
 Changed
 -------
+- `orix.quaternion.Quaternion` now relies on `numpy-quaternion <https://quaternion.readthedocs.io/en/latest/>`_
+  for quaternion conjugation, quaternion-quaternion and quaternion-vector multiplication,
+  and quaternion-quaternion and quaternion-vector outer products.
 - Rounding in functions is now set consistently at 12 dp, eg. `Object3d.unique` and 
   `Rotation.unique`.